% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{btw}
\alias{btw}
\title{If x is between left and right}
\usage{
btw(x, lower, upper)
}
\description{
If x is between left and right
}
\examples{
df <- data.frame(
  a = 1:5,
  lower = 2,
  upper = 4
)

dplyr::filter(df, btw(a, lower, upper))

}
