% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_pk}
\alias{check_pk}
\title{Check if a data.frame is uniquely identified by a given set of columns}
\usage{
check_pk(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{columns}
}
\description{
Check if a data.frame is uniquely identified by a given set of columns
}
\examples{
x <- expand.grid(a = 1:2, b = letters[1:3])
x

check_pk(x, a)
check_pk(x, a, b)

}
