% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{clip_cols}
\alias{clip_cols}
\title{Write column names of a data.frame to a clipboard}
\usage{
clip_cols(data, collapse = "\\n")
}
\arguments{
\item{data}{data.frame}
}
\description{
Write column names of a data.frame to a clipboard
}
\examples{
clip_cols(iris)
# Ctrl + V
# Sepal.Length
# Sepal.Width
# Petal.Length
# Petal.Width
# Species

clip_cols(iris, ", ")
# Paste the clipboard
# Sepal.Length, Sepal.Width, Petal.Length, Petal.Width, Species

}
