% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_duplicated}
\alias{find_duplicated}
\title{Find duplicated rows in data.frame}
\usage{
find_duplicated(data, ...)
}
\description{
Find duplicated rows in data.frame
}
\examples{
find_duplicated(mtcars, cyl)
}
