% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%btw\%}
\alias{\%btw\%}
\title{If x is with in a given range}
\usage{
x \%btw\% range
}
\description{
If x is with in a given range
}
\examples{
c(1, 1) \%btw\% list(c(0, 0), c(0, 1))
dplyr::filter(data.frame(x = 1:5, lw = 2, up = 4), x \%btw\% list(lw, up))

}
