% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_try_error}
\alias{is_try_error}
\title{Is an object the class of "try-error"?}
\usage{
is_try_error(x)
}
\description{
Is an object the class of "try-error"?
}
\examples{
x <- try(stop())
is_try_error(x)

}
