% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm.R
\name{params_sum_lnorm}
\alias{params_sum_lnorm}
\title{Compute mean and variance of the sum of log-normal distribution}
\usage{
params_sum_lnorm(mu, sigma)
}
\arguments{
\item{mu}{a vector}

\item{sigma}{a vector}
}
\value{
a vector of mu and sigma parameter
}
\description{
Compute the mean and variance of the sum of log-normal distributions based on Fenton-Wilkinson method
}
