% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnorm.R
\name{params_to_moments}
\alias{params_to_moments}
\title{Compute mean and variance of log-normal distribution}
\usage{
params_to_moments(mu, sigma)
}
\arguments{
\item{mu}{a vector}

\item{sigma}{a vector}
}
\value{
a matrix of mean and variance parameter
}
\description{
Compute mean and varianve of LogNormal(mu, sigma)
}
