% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws_s3_redshift.R
\name{replace_var}
\alias{replace_var}
\title{Replace a certain character in a text}
\usage{
replace_var(txt, var_sep = "?", var_list, ...)
}
\arguments{
\item{txt}{text}

\item{var_sep}{a character to separete a variable from a context. Default "?" as, "?var?"}

\item{var_list}{a list of variable names and characters to replace variables}
}
\description{
Replace a certain character in a text
}
\examples{
sql <- "
select
  user_id,
  names
from user_master
where 1=1
  and user_id = ?user_id?
;
"
cat(sql)
cat(replace_var(sql, user_id = 123))
}
