% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{view_in_excel}
\alias{view_in_excel}
\title{Create and open Excel workbook}
\usage{
view_in_excel(data, open_excel = TRUE)
}
\arguments{
\item{data}{data.frame}

\item{open_excel}{open workbook in Excel?}
}
\value{
workbook generated by openxlsx::createWorkbook
}
\description{
Create and open Excel workbook
}
\examples{
view_in_excel(iris) # Excel opens

}
