% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipboard.R
\name{write_to_clip}
\alias{write_to_clip}
\title{Write data.frame to a clipboard}
\usage{
write_to_clip(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{options for write.table}
}
\description{
Write data.frame to a clipboard
}
